Encryption
We will use so called one-time pad encryption, where client and server exchange list of keys and use one key for encryption only once. After a key has been used it must be discarded. You might question the exchange of keys over plain text TCP and you'd
be right to do so. This is an oversimplified example and should be used only to portray  one-time pad encryption as a example.